refactor: reorganize tests, reduce repetition
