+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testMatch: ['**/tests/**/*.test.?(m)js'],
+  moduleFileExtensions: ['js', 'mjs', 'json'],
+};
